backend/pyproject.toml
[project]
name = "ai-writing-detector"
version = "0.1.0"
description = "Rule-based AI writing detection API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "nltk>=3.9.0",
    "wordfreq>=3.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
fail_under = 85
show_missing = true
